@@ -14,14 +14,9 @@
         }
     ],
     "require": {
-        "php": "^8.1",
-        "openai-php/client": "^0.10.3",
-        "guzzlehttp/guzzle": "^7.0",
-        "illuminate/contracts": "^10.0||^11.0||^12.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^10.0",
-        "orchestra/testbench": "^8.0|^9.0"
+        "php": "^8.3",
+        "guzzlehttp/guzzle": "^7.10",
+        "openai-php/client": "^0.16.0"
     },
     "autoload": {
         "psr-4": {
@@ -30,11 +25,6 @@
         "files": [
             "src/helpers.php"
         ]
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "OpenAI\\Agents\\Tests\\": "tests/"
-        }
     },
     "extra": {
         "laravel": {
@@ -48,8 +38,16 @@
         }
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true,
+            "php-http/discovery": true
+        }
     },
     "minimum-stability": "dev",
-    "prefer-stable": true
+    "prefer-stable": true,
+    "require-dev": {
+        "orchestra/testbench": "^10.6",
+        "pestphp/pest": "^4.0"
+    }
 }
